requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
sqlalchemy>=1.4.0,<1.5.0
aiosqlite>=0.17.0,<0.18.0
python-dotenv>=0.19.0,<0.20.0
pydantic>=1.8.0,<2.0.0
jinja2>=3.0.0,<3.1.0
python-multipart>=0.0.5,<0.1.0 